+++ packages/daemon/src/v3/node/index.ts
@@ -1,3 +1,3 @@
 export { ESTOC_DIR, nodeHost, type NodeHostOptions } from "./host.js";
 export { serveDaemon, type ServeOptions, type Served } from "./serve.js";
-export { SOCKET_FILE, TOKEN_FILE, exitOnSignal, installedApp, runDaemon, type RunOptions } from "./run.js";
+export { SOCKET_FILE, TOKEN_FILE, exitOnSignal, installedApp, runDaemon, vaultDir, type RunOptions } from "./run.js";